Refactor flash summary logic
